function_app.py
import logging
import azure.functions as func

from main import check_for_new_products

app = func.FunctionApp()

@app.schedule(schedule="0 */1 * * * *", arg_name="myTimer", run_on_startup=True,
              use_monitor=False) 
def timer_trigger(myTimer: func.TimerRequest) -> None:
    if myTimer.past_due:
        logging.info('The timer is past due!')

    check_for_new_products()

    logging.info('Python timer trigger function executed.')